requirements.txt
Flask
mysql-connector-python
requests
Werkzeug
flask-paginate